Cleanup

+++ src/main/java/lt/vytzab/initiator/Order.java
@@ -7,7 +7,7 @@ public class Order implements Cloneable {
     private String symbol = null;
     private int quantity = 0;
     private int open = 0;
-    private double executed = 0;
+    private int executed = 0;
     private OrderSide side = OrderSide.BUY;
     private OrderType type = OrderType.MARKET;
     private OrderTIF tif = OrderTIF.DAY;
@@ -36,7 +36,8 @@ public class Order implements Cloneable {
             order.setOriginalID(getID());
             order.setID(order.generateID());
             return order;
-        } catch (CloneNotSupportedException e) {}
+        } catch (CloneNotSupportedException e) {
+        }
         return null;
     }
 
@@ -76,11 +77,13 @@ public class Order implements Cloneable {
         this.open = open;
     }
 
-    public double getExecuted() {
+    public int getExecuted() {
         return executed;
     }
 
-    public void setExecuted(double executed) { this.executed = executed; }
+    public void setExecuted(int executed) {
+        this.executed = executed;
+    }
 
     public OrderSide getSide() {
         return side;
